Added UpdateEntity Method, TransactionId => TransactionID

+++ UseCases/TransactionItems/UpsertTransactionItem/UpsertTransactionItemRequest.cs
@@ -10,7 +10,7 @@ public class UpsertTransactionItemRequest
 
     public int Quantity { get; set; }
 
-    public int PricePerItem { get; set; }
+    public decimal PricePerItem { get; set; }
 
     public decimal LineGross { get; set; }
